FIX: Enhanced Error Context + FEAT: 7-Day Hits Chart
**Code Review Fixes v8.7:**

**1. UI Config: Bessere Error Messages (Overall Comment #1)**
   - PROBLEM: assertInRange() verlor Konstanten-Namen im Error
   - VORHER: "MIN_DURATION_SLIDER_MAX (2000) must be between 100 and 10000"
   - NACHHER: "MIN_DURATION_SLIDER_MAX (2000) must be between
              MIN_SESSION_DURATION_MS (100) and MAX_SESSION_DURATION_MS (10000)"
   - Vorteil: Besserer Debugging-Kontext mit Konstanten-Namen

   **Anpassung assertInRange Signatur:**
   ```javascript
   // Vorher:
   assertInRange(name, value, min, max)

   // Nachher:
   assertInRange(name, value, minName, min, maxName, max)
   ```

**2. ESP32: Explizite Casts (Overall Comment #2)**
   - PROBLEM: Implizite Type Conversion könnte Überraschungen bergen
   - LÖSUNG: Explizite (long) casts in snprintf
   ```cpp
   snprintf(..., (long)minValue, (long)maxValue);  // Explicit casts
   ```
   - Vorteil: Klarheit + verhindert implicit conversion issues

**3. UI Config: const Arrow Functions (Overall Comment #3)**
   - PROBLEM: function-Deklarationen für module-level validation
   - LÖSUNG: const arrow functions
   ```javascript
   // Vorher:
   function assertInRange(...) { ... }

   // Nachher:
   const assertInRange = (...) => { ... };
   ```
   - Vorteil: Immutable + klarere Intent für module-level helpers

**NEU FEATURE: 7-Tage Hits Liniendiagramm**

**Location:** AnalyticsView - Erstes Widget nach Header

**Features:**
- **Bar Chart** der letzten 7 Tage (Mo-So)
- **Farb-Codierung**:
  * Heute: Emerald (grün)
  * Andere Tage mit Hits: Blue
  * Tage ohne Hits: Grau (zinc-800)
- **Y-Achse** mit dynamischer Skalierung (0, max/2, max)
- **Interaktiv**: Hover zeigt Tooltip mit Tag + Anzahl
- **3 Stats-Karten**:
  1. Ø pro Tag (Durchschnitt)
  2. Höchster Tag (Maximum)
  3. Heute (aktueller Stand)
- **Gesamt-Counter** im Header

**Implementierung:**
```javascript
const last7DaysData = useMemo(() => {
  // Generiert Array der letzten 7 Tage
  // Matched mit historyData
  // Returns: [{ date, day, count }, ...]
}, [historyData]);
```

**Visual Design:**
- Gradient Bars (from-x-600 to-x-400)
- Rounded tops (rounded-t-lg)
- Fixed height (140px) mit dynamischem fill %
- Responsive gap-2 zwischen Bars
- Dark theme (zinc-900/950)

**Data Flow:**
historyData → last7DaysData → Bar heights → Visual representation

**Addressed Comments:**
- Overall 1: Konstanten-Namen in Errors ✓
- Overall 2: Explizite Casts ✓
- Overall 3: const arrow functions ✓

**User Request:**
"statistiktab bitte um liniendiagramm für die hits der letzten 7 tage"

+++ esp32-firmware/highscore-sensor-c3.ino
@@ -248,23 +248,24 @@ void readBatteryVoltage() {
 
 // ===== API HELPER FUNCTIONS =====
 /**
- * **FIX v8.6**: Helper für Range-Error-Messages (reusable + testable)
+ * **FIX v8.6 + v8.7**: Helper für Range-Error-Messages (reusable + testable)
  * Baut eine JSON-Error-Message für ungültige Session-Dauer-Ranges
  *
  * @param fieldName Name des Feldes (z.B. "minSessionDuration")
  * @param buffer Output-Buffer für JSON
  * @param bufferSize Größe des Buffers
- * @param minValue Minimaler erlaubter Wert
- * @param maxValue Maximaler erlaubter Wert
+ * @param minValue Minimaler erlaubter Wert (explizit als long gecastet für Klarheit)
+ * @param maxValue Maximaler erlaubter Wert (explizit als long gecastet für Klarheit)
  */
 void buildRangeError(const char* fieldName, char* buffer, size_t bufferSize, long minValue, long maxValue) {
+  // **FIX v8.7**: Explizite Casts für Klarheit (verhindert implicit conversion Überraschungen)
   snprintf(
     buffer,
     bufferSize,
     "{\"error\":\"%s muss zwischen %ld-%ldms liegen\"}",
     fieldName,
-    minValue,
-    maxValue
+    (long)minValue,  // Explicit cast
+    (long)maxValue   // Explicit cast
   );
 }
